requirements.txt
absl-py==0.6.1
astor==0.7.1
beautifulsoup4==4.8.2
bs4==0.0.1
cycler==0.10.0
Django==2.2.10
ffmpeg==1.4
gast==0.2.0
grpcio==1.17.0
kiwisolver==1.1.0
Markdown==3.0.1
matplotlib==3.0.3
numpy==1.14.5
pandas==0.25.3
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.2
scipy==1.3.0
six==1.11.0
soupsieve==1.9.5
sqlparse==0.3.0
tensorboard==1.10.0
termcolor==1.1.0
virtualenv==16.1.0
Werkzeug==1.0.0
